CMSIS-DSP: Ensure correlation array index is signed.
This attempts to ensures that even on a 64bit system, the array access
will be treated as a negative number, not a large unsigned offset ending
up reading from the middle of nowhere.

+++ CMSIS/DSP/Source/FilteringFunctions/arm_correlate_q15.c
@@ -882,7 +882,7 @@ void arm_correlate_q15(
       if (((i - j) < srcBLen) && (j < srcALen))
       {
         /* z[i] += x[i-j] * y[j] */
-        sum += ((q31_t) pIn1[j] * pIn2[-((int32_t) i - j)]);
+        sum += ((q31_t) pIn1[j] * pIn2[-((int32_t) i - (int32_t) j)]);
       }
     }
